+++ src/GammonX/GammonX.Server/Services/ICancellationTokenService.cs
@@ -1,0 +1,59 @@
+﻿namespace GammonX.Server.Services
+{
+    /// <summary>
+    /// Provides the capability to mmanage cancellation tokens application wide.
+    /// </summary>
+    public interface ICancellationTokenService
+    {
+        /// <summary>
+        /// Stores a cancellation token source for a given key.
+        /// </summary>
+        /// <remarks>
+        /// If a token source already exists for the given key, it is cancelled and overwritten.
+        /// </remarks>
+        /// <param name="key">Timer key.</param>
+        /// <param name="cts">Token to cancel the timer.</param>
+        void Store(CancellationTokenKey key, CancellationTokenSource cts);
+
+        /// <summary>
+        /// Tries to get a cancellation token source for a given key.
+        /// </summary>
+        /// <param name="key">Key to search for.</param>
+        /// <param name="cts">Token result.</param>
+        /// <returns>Returns true if one is found, otherwise null.</returns>
+        bool TryGet(CancellationTokenKey key, out CancellationTokenSource? cts);
+
+        /// <summary>
+        /// Cancels the given token source by the specified <paramref name="key"/>.
+        /// </summary>
+        /// <param name="key">Key to search for.</param>
+        void Cancel(CancellationTokenKey key);
+
+        /// <summary>
+        /// Cancels all token sources associated with the given player id.
+        /// </summary>
+        /// <param name="playerId"></param>
+        void CancelForPlayer(Guid playerId);
+
+        /// <summary>
+        /// Cancels all token sources associated with the given match id.
+        /// </summary>
+        /// <param name="matchid"></param>
+        void CancelForMatch(Guid matchid);
+    }
+
+    public readonly struct CancellationTokenKey
+    {
+        public CancellationTokenCategory Category { get; init; }
+
+        public Guid MatchId { get; init; }
+
+        public Guid PlayerId { get; init; }
+    }
+
+    public enum CancellationTokenCategory
+    {
+        Turn = 0,
+        Disconnect = 1,
+    }
+}